% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateDMRs.R
\name{annotateDMRs}
\alias{annotateDMRs}
\title{Output annotated DMRs in text and gff3 format, and DMR count table}
\usage{
annotateDMRs(annotation, gff.files, if.gff3, input.dir, out.dir)
}
\arguments{
\item{annotation}{Annotation terms used to annotate DMRs. (char)}

\item{gff.files}{Paths to the gff3 files can be supplied as a vector; 
corresponding to the annotation terms. (char)}

\item{if.gff3}{A logical specifying whether output annotated files in 
gff3. (logical)}

\item{input.dir}{Path to the input directory containing (non)/filtered 
DMR matrix/matrices. (char)}

\item{out.dir}{Path to the output directory. (char)}
}
\value{
Output files containing annotated DMRs and DMR counts table
}
\description{
Output annotated DMRs in text and gff3 format, and DMR count table
}
\examples{
## any file with chromosome, start and end can be annotated
ind <- file.path(tempdir(), "jDMRgrid_toannotate")
dir.create(ind, showWarnings = FALSE)
data.table::fwrite(
    data.frame(seqnames = 1L, start = c(2131L, 8736L), end = c(3630L, 10237L)),
    file.path(ind, "CG_regions.txt"), sep = "\t")

gff <- system.file("extdata/toyData", "TAIR10_promoters.gff3", package = "jDMRgrid")
outd <- file.path(tempdir(), "jDMRgrid_annot")
annotateDMRs(annotation = "promoters", gff.files = gff, if.gff3 = FALSE,
             input.dir = ind, out.dir = outd)
list.files(outd)
}
