% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDMRmatrix.R
\name{filterEpiMAF}
\alias{filterEpiMAF}
\title{Application of filtering using epiMAF criterion}
\usage{
filterEpiMAF(status.collect, rc.methlevel.collect, epiMAF)
}
\arguments{
\item{status.collect}{DataFrame for state-calls. (DataFrame object)}

\item{rc.methlevel.collect}{DataFrame rc-methylation levels. 
(DataFrame object)}

\item{epiMAF}{Threshold to filter for minor epi-allele frequency. (num)}
}
\value{
List of data frames for rc-methylation level and state-calls using 
        minor epi-allele frequency criterion.
}
\description{
Application of filtering using epiMAF criterion
}
\examples{
## state calls and methylation levels for one context
co <- data.frame(seqnames = 1L, start = c(1L, 201L, 401L), end = c(200L, 400L, 600L))
sc <- data.frame(co, s1 = c(1, 0, 1), s2 = c(1, 0, 1),
                     s3 = c(1, 0, 1), s4 = c(0, 0, 1))
rc <- data.frame(co, s1 = c(0.9, 0.1, 0.8), s2 = c(0.9, 0.1, 0.8),
                     s3 = c(0.9, 0.1, 0.8), s4 = c(0.1, 0.1, 0.8))

## keep regions where the rarer state is carried by fewer than 30 percent of samples
res <- filterEpiMAF(sc, rc, epiMAF = 0.3)
res[[1]]
}
