% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDMRmatrix.R
\name{filterReplicateConsensus}
\alias{filterReplicateConsensus}
\title{Application of filtering using replicate consensus}
\usage{
filterReplicateConsensus(
  status.collect,
  rc.methlevel.collect,
  replicate.consensus,
  diff.ct
)
}
\arguments{
\item{status.collect}{DataFrame for state-calls. (DataFrame object)}

\item{rc.methlevel.collect}{DataFrame rc-methylation levels. 
(DataFrame object)}

\item{replicate.consensus}{Percentage of concordance in state calls. (num)}

\item{diff.ct}{Allowed difference between control and treatment groups. (num)}
}
\value{
List of data frames for rc-methylation level and state-calls using 
        replicate-consensus criterion.
}
\description{
Application of filtering using replicate consensus
}
\examples{
## two groups of two replicates, named <sample>_<replicate>
co <- data.frame(seqnames = 1L, start = c(1L, 201L, 401L), end = c(200L, 400L, 600L))
sc <- data.frame(co, WT_rep1 = c(1, 0, 1), WT_rep2 = c(1, 0, 1),
                     mut_rep1 = c(0, 0, 1), mut_rep2 = c(0, 0, 1))
rc <- data.frame(co, WT_rep1 = c(0.9, 0.1, 0.9), WT_rep2 = c(0.9, 0.1, 0.9),
                     mut_rep1 = c(0.1, 0.1, 0.9), mut_rep2 = c(0.1, 0.1, 0.9))

## replicates must agree within each group, and the groups must differ
res <- filterReplicateConsensus(sc, rc, replicate.consensus = 1, diff.ct = 0.5)
res[[1]]
}
