\docType{data}
\name{listFiles1}
\alias{listFiles1}
\title{samplelist file for population-based data}
\format{A \code{\link[S4Vectors]{DataFrame-class}} object.}
\description{
A \code{\link[S4Vectors]{DataFrame-class}} object for demonstration purposes in 
examples of package jDMRgrid. 
The object contains a samplelist file for population-based data
}
\value{
samplelist file for population-based data
}
\examples{
data(listFiles1)
print(listFiles1)
}