\docType{data}
\name{listFiles2}
\alias{listFiles2}
\title{samplelist file for replicate-based data}
\format{A \code{\link[S4Vectors]{DataFrame-class}} object.}
\description{
A \code{\link[S4Vectors]{DataFrame-class}} object for demonstration purposes in 
examples of package jDMRgrid. 
The object contains a samplelist file for replicate-based data
}
\value{
samplelist file for replicate-based data
}
\examples{
data(listFiles2)
print(listFiles2)
}