% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDMRmatrix.R
\name{makeDMRmatrix}
\alias{makeDMRmatrix}
\title{Builds a DMR matrix for all samples
This function generates a binary matrix, a matrix of
recalibrated methylation levels and posterior probabilities for all samples.}
\usage{
makeDMRmatrix(
  contexts = c("CG", "CHG", "CHH"),
  postMax.out = TRUE,
  samplelist,
  input.dir,
  out.dir,
  include.intermediate = FALSE,
  if.Bismark = FALSE
)
}
\arguments{
\item{contexts}{Vector of cytosine contexts; default c('CG','CHG','CHH').
(char)}

\item{postMax.out}{Logical: also write the matrix of posterior probabilities
behind each state call. Needed by filterDMRmatrix(min.posterior=)
and by rankDMRs(). Costs roughly 35-40 percent more output.
(logical; default as TRUE)}

\item{samplelist}{DataFrame object containing information about
file, sample, replicate and group. (DataFrame object)}

\item{input.dir}{Path to the directory holding the region-call files written
by runjDMRgrid(). (char)}

\item{out.dir}{Path to the output directory. (char)}

\item{include.intermediate}{Logical if intermediate calls should be used;
default as FALSE. (logical)}

\item{if.Bismark}{Whether the inputs were Bismark CX reports; controls how sample
names are derived from file paths. (logical)}
}
\description{
Builds a DMR matrix for all samples
This function generates a binary matrix, a matrix of
recalibrated methylation levels and posterior probabilities for all samples.
}
\examples{
sheet <- get(load(system.file("data", "listFiles2.RData", package = "jDMRgrid")))
sheet$file <- system.file("extdata", sheet$file, package = "jDMRgrid")

grid <- file.path(tempdir(), "jDMRgrid_grid2")
mat  <- file.path(tempdir(), "jDMRgrid_matrix2")
runjDMRgrid(out.dir = grid, window = 200, step = 50, samplelist = sheet,
            contexts = "CG", min.C = 10, min.C.type = "percentile",
            mincov = 0, include.intermediate = FALSE, runName = "example")
makeDMRmatrix(contexts = "CG", samplelist = sheet,
              input.dir = grid, out.dir = mat, include.intermediate = FALSE)
list.files(mat)
}
