% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runjDMR.R
\name{makeMethimpute_foreach}
\alias{makeMethimpute_foreach}
\title{Run makeMethimpute using future_apply parallel method}
\usage{
makeMethimpute_foreach(
  out.samplelist,
  merge_list,
  include.intermediate,
  out.dir,
  mincov,
  numCores,
  if.Bismark,
  cyt.pos.all
)
}
\arguments{
\item{out.samplelist}{DataFrame consisting of: contexts, methImpute files,
sample names (from methImpute files) and IDs 
corresponding to the binned genome. (DataFrame object)}

\item{merge_list}{List of DataFrames with binned regions; for each context
separately (list of DataFrame objects)}

\item{include.intermediate}{Logical if intermediate calls should be used;
default as FALSE. (logical)}

\item{out.dir}{Path to the output directory. (char)}

\item{mincov}{Minimum read coverage; default as 0. (num; between 0 and 1)}

\item{numCores}{Number of cores to be used if for_each method should be 
performed; default as NULL. (num)}

\item{if.Bismark}{Logical if Bismark inputs (CX reports in txt format)
are used. Default as FALSE. (logical)}

\item{cyt.pos.all}{GRanges object from extractCytosinesFromFASTA. (GRanges)}
}
\value{
Methylome for regions taken out grid genome from non/sliding window
        approach
}
\description{
Run makeMethimpute using future_apply parallel method
}
