% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitGroups.R
\name{splitGroups}
\alias{splitGroups}
\title{Split into control/treatment groups according to the sample and replicates}
\usage{
splitGroups(
  samplelist,
  postMax.out = TRUE,
  contexts = c("CG", "CHG", "CHH"),
  input.dir,
  out.dir
)
}
\arguments{
\item{samplelist}{DataFrame object containing information about
file, sample, replicate and group. (DataFrame object)}

\item{postMax.out}{Logical if postMax matrices should be included. (logical)}

\item{contexts}{Vector of cytosine contexts; default as c('CG','CHG','CHH').
(char)}

\item{input.dir}{Path to the input directory with DMR matrices. (char)}

\item{out.dir}{Path to the output directory for DMR matrices. (char)}
}
\value{
DMR matrices split by sample/treatment groups, saved in the same
        directory with DMR matrices.
}
\description{
Split into control/treatment groups according to the sample and replicates
}
\examples{
## a small DMR matrix, written to a temporary directory
d <- file.path(tempdir(), "jDMRgrid_example")
dir.create(d, showWarnings = FALSE)
co  <- data.frame(seqnames = 1L, start = c(1L, 201L, 401L), end = c(200L, 400L, 600L))
smp <- c("WT_rep1", "WT_rep2", "mutant1_rep1", "mutant1_rep2")
st  <- rbind(c(1, 1, 0, 0), c(0, 0, 1, 1), c(1, 1, 1, 1))
mk  <- function(nm, M) {
    x <- data.frame(co, M)
    names(x) <- c("seqnames", "start", "end", smp)
    data.table::fwrite(x, file.path(d, nm), sep = "\t")
}
mk("CG_StateCalls.txt", st)
mk("CG_rcMethlvl.txt", st * 0.9)
mk("CG_postMax.txt", matrix(0.99, 3, 4))

sheet <- data.frame(
    file = NA_character_, sample = rep(c("WT", "mutant1"), each = 2),
    replicate = rep(c("rep1", "rep2"), 2),
    group = rep(c("control", "treatment1"), each = 2), stringsAsFactors = FALSE)

## one matrix per pairwise comparison
splitGroups(samplelist = sheet, contexts = "CG", input.dir = d, out.dir = d)
list.files(d, pattern = "WT_mutant1")
}
