% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDMRmatrix.R
\name{filterDMRmatrix}
\alias{filterDMRmatrix}
\title{filter DMR matrix}
\usage{
filterDMRmatrix(
  epiMAF.cutoff = NULL,
  replicate.consensus = NULL,
  data.dir,
  samplefiles
  mergingBins = TRUE
)
}
\arguments{
\item{epiMAF.cutoff}{Numeric threshold to filter for minor epi-allele frequency. Applicable for population level data. By default this option is set to NULL. (NULL or numeric value between 0 and 1)}

\item{replicate.consensus}{Numeric threshold as the percentage of concordance in methylation states among samples with multiple replicates. Applicable for control/treatment data. By default this option is set to NULL. (NULL or numeric value between 0 and 1)}

\item{data.dir}{Path to the directory containing DMR matrix files. Looks for files with suffix("_StateCalls.txt" and "_rcMethlvl.txt"). (character)}

\item{samplefiles}{Path to the text file containing path to samples and sample names. For control/treatment data an additional column specifying the replicates is required. (character)}

\item{mergingBins}{Logical argument if merging consecutive bins having the same stateCalls should be performed. (logical)}
}
\description{
Filters non-polymorphic patterns by default.
}
